requirements.txt
fastapi
uvicorn
diffusers
transformers
accelerate
safetensors
torch
invisible-watermark
pillow
jinja2
matplotlib
